# src/agent/Hercules.WebApi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Hercules.WebApi | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/config": {
      "get": {
        "tags": [
          "Config"
        ],
        "operationId": "GetConfig",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Config"
        ],
        "operationId": "UpdateConfig",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonElement"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Config"
        ],
        "operationId": "PatchConfig",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonElement"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/approvals/pending": {
      "get": {
        "tags": [
          "Approvals"
        ],
        "operationId": "ListPendingApprovals",
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/approvals/{id}": {
      "get": {
        "tags": [
          "Approvals"
        ],
        "operationId": "GetApproval",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/approvals/{id}/approve": {
      "post": {
        "tags": [
          "Approvals"
        ],
        "operationId": "ApproveTool",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/approvals/{id}/deny": {
      "post": {
        "tags": [
          "Approvals"
        ],
        "operationId": "DenyTool",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/approvals": {
      "get": {
        "tags": [
          "Approvals"
        ],
        "operationId": "ListAllApprovals",
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/context/summary": {
      "get": {
        "tags": [
          "Context"
        ],
        "operationId": "ContextSummary",
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxTokens",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/context/distill": {
      "post": {
        "tags": [
          "Context"
        ],
        "operationId": "ContextDistill",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DistillRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/context/trace/compress": {
      "post": {
        "tags": [
          "Context"
        ],
        "operationId": "ContextTraceCompress",
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ToolTraceEntry"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/cache/invalidate": {
      "post": {
        "tags": [
          "Cache"
        ],
        "operationId": "CacheInvalidate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CacheInvalidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CacheInvalidateRequest": {
        "type": "object",
        "properties": {
          "all": {
            "type": "boolean",
            "description": "Invalidate all cache classes."
          },
          "class": {
            "type": [
              "null",
              "string"
            ],
            "description": "Cache class to invalidate (e.g. \"Embedding\", \"RoutingDecision\")."
          },
          "key": {
            "type": [
              "null",
              "string"
            ],
            "description": "Specific key to invalidate (requires 'class')."
          },
          "pattern": {
            "type": [
              "null",
              "string"
            ],
            "description": "Key pattern to invalidate (requires 'class')."
          }
        },
        "description": "Request body for cache invalidation."
      },
...
      "DistillRequest": {
        "required": [
          "sessionId",
          "mode"
        ],
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string"
          },
          "mode": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "Request body для `POST /api/context/distill` (task_102)."
      },
...
      "JsonElement": { },
...
      "ToolTraceEntry": {
        "required": [
          "toolName",
          "input",
          "output",
          "durationMs",
          "timestamp",
          "success"
        ],
        "type": "object",
        "properties": {
          "toolName": {
            "type": "string"
          },
          "input": {
            "type": "string"
          },
          "output": {
            "type": "string"
          },
          "durationMs": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}